+++ tools/packager/main.cpp
@@ -43,10 +43,10 @@ void packFile(std::ofstream & out, const std::string & file) {
 	std::string filename = i6engine::utils::split(file, "/").back();
 	filename = i6engine::utils::split(filename, "\\").back();
 
-	uint32_t filenameLength = filename.length();
+	uint32_t filenameLength = uint32_t(filename.length());
 	clockUtils::compression::Compression<clockUtils::compression::algorithm::HuffmanGeneric> compressor;
 	std::string compressed = compressor.compress(std::string(buffer, size_t(size)));
-	uint32_t length = compressed.length();
+	uint32_t length = uint32_t(compressed.length());
 	out.write(reinterpret_cast<char *>(&length), sizeof(uint32_t));
 	out.write(reinterpret_cast<char *>(&filenameLength), sizeof(uint32_t));
 	out.write(filename.c_str(), filename.size());
